implement window centering
This is a custom centering patch. The centering patch on the suckless
website is extremely slow when reading large files as it needs to read
all items, then loop though and determine the string lengths before
displaying a window. This patch centers the window, but the width must
be manually preset with the `-z` flag.